Image and Chatbot/requirements.txt
streamlit
google-generativeai
python-dotenv
ipython==8.10.0
